front/pyproject.toml
[project]
name = "front"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "dishka>=1.5.3",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "markdown>=3.8",
    "nats-py>=2.10.0",
    "orjson>=3.10.18",
    "python-multipart>=0.0.20",
    "python-socketio[asgi]>=5.13.0",
    "pyyaml>=6.0.2",
    "structlog>=25.2.0",
    "uvicorn>=0.34.2",
]


[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"
show-fixes = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "S101",   # Use of `assert` detected
    "CPY001", # Missing copyright notice at top of file,
    "DOC501", # Raised exception missing from docstring
    "DOC201", # `return` is not documented in docstring
    "D100",   # Missing docstring in public module
    "D101",   # Missing docstring in public class
    "D106",   # Missing docstring in public nested class
    "D107",   # Missing docstring in `__init__`
    "D203",   # 1 blank line required before class docstring
    "D211",   # No blank lines allowed before function docstring
    "D213",   # Multi-line docstring summary should start at the second line
    "DOC402", # `yield` is not documented in docstring
    "COM812", # Trailing comma missing, conflicts with formatter
    "RUF001", # String contains ambiguous `с` (CYRILLIC SMALL LETTER ES). Did you mean `c` (LATIN SMALL LETTER C)?
    "RUF002", # Docstring contains ambiguous `с` (CYRILLIC SMALL LETTER ES). Did you mean `c` (LATIN SMALL LETTER C)?
    "RUF003", # ambiguous `с` (CYRILLIC SMALL LETTER ES). Did you mean `c` (LATIN SMALL LETTER C)?
    "TC001",  # Move application import into a type-checking block
    "TC002",  # Move third-party import
    "TC003",  # Move standard library import into a type-checking bloc
    "TRY400", # Use `logging.exception` instead of `logging.error`
    "PLR0917",# Too many arguments in function definition
    "PLR0913",# Too many arguments in function definition > 5
    "FIX002", # Line contains TODO, consider resolving the issue
    "TD002",  # Missing author in TODO;
    "TD003",  # Missing issue link on the line following this TODO
    "ISC001", # The following rule may cause conflicts when used with the formatter: `ISC001`
    "FBT001", # Boolean-typed positional argument in function definition
    "FURB118",# Use `operator.itemgetter("name")` instead of defining a lambda
]
unfixable = [
    "B905",
]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = ["UP", "S608", "N999", "D"]
"*/di/*" = [
    "D102",    # Missing docstring in public method
    "PLR6301", # Method could be a function, class method, or static metho
]
"*/routes/*" = [
    "D301",    # Use `r"""` if any backslashes in a docstring
]
"*/tests/*" = [
    "ANN",     # Annotations required
    "D",       # Documentation required
    "PTH",     #
    "SLF001",  #
    "PLR6301", # Method could be a function, class method, or static method
    "S404",    # `subprocess` module is possibly insecure
    "PLR2004", # Magic value used in comparison
    "ARG001"   # Unused function argument
]
"*/main/*" = [
    "RUF029"   # Function is declared `async`, but doesn't `await` or use `async` features.
]
"*/__init__.py" = ["F401", "D104"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.isort]
length-sort = false

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
exclude = ["tests", "alembic"]
warn_unused_ignores = false

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
    "types-markdown>=3.8.0.20250415",
    "types-pyyaml>=6.0.12.20250402",
]
